// Qanat.Web/src/app/shared/generated/model/well-registration-grid-row-dto.ts
export class WellRegistrationGridRowDto { 
    WellRegistrationID?: number;
    GeographyID?: number;
    WellID?: number;
    WellName?: string;
    WellRegistrationStatusID?: number;
    ParcelID?: number;
    ParcelNumber?: string;
    SubmitDate?: string;
    ApprovalDate?: string;
    CreateUserName?: string;
    CreateUserEmail?: string;
    Latitude?: number;
    Longitude?: number;
    IrrigatedParcels?: Array<WellRegistrationGridRowIrrigatedParcelDto>;
    LandownerName?: string;
    LandownerBusinessName?: string;
    LandownerStreetAddress?: string;
    LandownerCity?: string;
    LandownerState?: string;
    LandownerZipCode?: string;
    LandownerPhone?: string;
    LandownerEmail?: string;
    OwnerOperatorName?: string;
    OwnerOperatorBusinessName?: string;
    OwnerOperatorStreetAddress?: string;
    OwnerOperatorCity?: string;
    OwnerOperatorState?: string;
    OwnerOperatorZipCode?: string;
    OwnerOperatorPhone?: string;
    OwnerOperatorEmail?: string;
    StateWCRNumber?: string;
    CountyWellPermitNumber?: string;
    DateDrilled?: string;
    AgriculturalWaterUse?: boolean;
    AgriculturalWaterUseDescription?: string;
    StockWateringWaterUse?: boolean;
    StockWateringWaterUseDescription?: string;
    DomesticWaterUse?: boolean;
    DomesticWaterUseDescription?: string;
    PublicMunicipalWaterUse?: boolean;
    PublicMunicipalWaterUseDescription?: string;
    PrivateMunicipalWaterUse?: boolean;
    PrivateMunicipalWaterUseDescription?: string;
    OtherWaterUse?: boolean;
    OtherWaterUseDescription?: string;
    WellRegistrationMetadatum?: WellRegistrationMetadatumSimpleDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}